% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBPSMain.R
\name{balance.CBPS}
\alias{balance.CBPS}
\title{Calculates the pre- and post-weighting difference in standardized means for covariate within each contrast}
\usage{
\method{balance}{CBPS}(object, ...)
}
\arguments{
\item{object}{A CBPS, npCBPS, or CBMSM object.}

\item{...}{Additional arguments to be passed to balance.}
}
\description{
Calculates the pre- and post-weighting difference in standardized means for covariate within each contrast
}
