% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBPSMain.R
\name{balance.CBPSContinuous}
\alias{balance.CBPSContinuous}
\title{Calculates the pre- and post-weighting correlations between each covariate and the T}
\usage{
\method{balance}{CBPSContinuous}(object, ...)
}
\description{
Calculates the pre- and post-weighting correlations between each covariate and the T
}
