\name{Blackwell}
\docType{data}
\alias{Blackwell}
\title{Blackwell Data for Covariate Balancing Propensity Score}
\description{This data set gives the outcomes a well as treatment assignments and covariates for the example from Blackwell (2013).}
\usage{Blackwell}
\format{A data frame consisting of 13 columns (including treatment assignment, time, and identifier vectors) and 570 observations.  }
\source{Data from the National Supported Work Study.  A benchmark matching dataset.  d.gone.neg is the treatment. d.gone.neg.l1, d.gone.neg.l2, and d.gone.neg.l3 are lagged treatment variables. camp.length, deminc, base.poll, base.und, and office covariates. year and time are alternative specifications of the time variable. demName is the identifier, and demprcnt is the outcome.}
\references{Blackwell, Matthew. (2013). A framework for dynamic causal inference in political science. American Journal of Political Science 57, 2, 504-619.}
\keyword{datasets}