% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBPSMain.R
\name{CBPS.fit}
\alias{CBPS.fit}
\title{CBPS.fit determines the proper routine (what kind of treatment) and calls the 
approporiate function.  It also pre- and post-processes the data}
\usage{
CBPS.fit(treat, X, baselineX, diffX, ATT, method, iterations, standardize,
  twostep, sample.weights = sample.weights, ...)
}
\description{
CBPS.fit determines the proper routine (what kind of treatment) and calls the 
approporiate function.  It also pre- and post-processes the data
}
