% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBPSMain.R
\name{plot.CBPSContinuous}
\alias{plot.CBPSContinuous}
\title{Plot the pre-and-post weighting correlations between X and T}
\usage{
\method{plot}{CBPSContinuous}(x, covars = NULL, silent = TRUE,
  boxplot = FALSE, ...)
}
\description{
Plot the pre-and-post weighting correlations between X and T
}
