% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBPSMain.R
\name{print.CBPS}
\alias{print.CBPS}
\title{Print coefficients and model fit statistics}
\usage{
\method{print}{CBPS}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
Print coefficients and model fit statistics
}
