% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npCBPS.R
\name{balance.npCBPS}
\alias{balance.npCBPS}
\title{Calls the appropriate balance function based on the number of treatments}
\usage{
\method{balance}{npCBPS}(object, ...)
}
\description{
Calls the appropriate balance function based on the number of treatments
}
