% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBPS-package.R
\docType{data}
\name{Blackwell}
\alias{Blackwell}
\title{Blackwell Data for Covariate Balancing Propensity Score}
\format{
A data frame consisting of 13 columns (including treatment
assignment, time, and identifier vectors) and 570 observations.
}
\source{
d.gone.neg is the treatment. d.gone.neg.l1, d.gone.neg.l2, and
d.gone.neg.l3 are lagged treatment variables. camp.length, deminc,
base.poll, base.und, and office covariates. year is the year of the
particular race, and time goes from the first measurement (time = 1) to the
election (time = 5). demName is the identifier, and demprcnt is the outcome.
}
\description{
This data set gives the outcomes a well as treatment assignments and
covariates for the example from Blackwell (2013).
}
\references{
Blackwell, Matthew. (2013). A framework for dynamic causal
inference in political science. American Journal of Political Science 57, 2,
504-619.
}
\keyword{datasets}
