% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npCBPS.R
\name{plot.npCBPS}
\alias{plot.npCBPS}
\title{Calls the appropriate plot function, based on the number of treatments}
\usage{
\method{plot}{npCBPS}(x, covars = NULL, silent = TRUE, ...)
}
\description{
Calls the appropriate plot function, based on the number of treatments
}
