% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBPSMain.R
\name{print.CBPS}
\alias{print.CBPS}
\title{Print coefficients and model fit statistics}
\usage{
\method{print}{CBPS}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class \dQuote{CBPS} or \dQuote{npCBPS}, usually, a result of a call to \code{CBPS} or \code{npCBPS}.}

\item{digits}{the number of digits to keep for the numerical quantities.}

\item{...}{Additional arguments to be passed to summary.}
}
\description{
Print coefficients and model fit statistics
}
