% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_vcov.R
\name{vcov_outcome.CBPSContinuous}
\alias{vcov_outcome.CBPSContinuous}
\title{vcov_outcome}
\usage{
\method{vcov_outcome}{CBPSContinuous}(object, Y, Z, delta, tol = 10^(-5),
  lambda = 0.01)
}
\description{
vcov_outcome
}
